Dependencies and File Structure (#4)
Now dependencies are downloaded on the startup, solving the need to create a build script to make the plugin usable from the repository.
File structure was altered to follow the common practice, now this repo can be extracted into the pykrita directory and it should work as any other plugin.
Previously I didn't concern in making the file structure usable because I didn't want to put the .git directly on the pykrita directory and I was planning to make a build script anyway to download the dependencies.
Thanks to [twoda5h](https://github.com/twoda5h) for trying to fix the file structure earlier and sorry for not looking the PR for a month